add test-utils and rename util folder to utils
