+++ src/components/AmountInput.tsx
@@ -3,7 +3,11 @@ import { ChangeEvent, FC, useContext, useState } from "react";
 import type { FieldError, UseFormRegisterReturn } from "react-hook-form";
 import { useTranslation } from "react-i18next";
 import { AppContext, Unit } from "../context/app-context";
-import { convertBtcToSat, convertSatToBtc, formatAmount } from "../util/format";
+import {
+  convertBtcToSat,
+  convertSatToBtc,
+  formatAmount,
+} from "../utils/format";
 
 export type Props = {
   amount?: number;
