+++ src/context/app-context.tsx
@@ -17,7 +17,7 @@ import {
   retrieveSettings,
   saveSettings,
   setWindowAlias,
-} from "../util";
+} from "../utils";
 import { SSEContext } from "./sse-context";
 
 interface AppContextType {
