+++ src/pages/Home/LightningCard.tsx
@@ -3,8 +3,8 @@ import { useTranslation } from "react-i18next";
 import LoadingBox from "../../components/LoadingBox";
 import { AppContext, Unit } from "../../context/app-context";
 import { SSEContext } from "../../context/sse-context";
-import { convertMSatToBtc, convertToString } from "../../util/format";
-import { checkPropsUndefined } from "../../util";
+import { convertMSatToBtc, convertToString } from "../../utils/format";
+import { checkPropsUndefined } from "../../utils";
 
 export const LightningCard: FC = () => {
   const { t } = useTranslation();
