+++ src/pages/Home/ReceiveModal/ReceiveModal.tsx
@@ -7,10 +7,10 @@ import { useTranslation } from "react-i18next";
 import Message from "../../../components/Message";
 import ModalDialog from "../../../layouts/ModalDialog";
 import { AppContext, Unit } from "../../../context/app-context";
-import { checkError } from "../../../util/checkError";
-import { convertBtcToSat, stringToNumber } from "../../../util/format";
-import { instance } from "../../../util/interceptor";
-import { MODAL_ROOT } from "../../../util";
+import { checkError } from "../../../utils/checkError";
+import { convertBtcToSat, stringToNumber } from "../../../utils/format";
+import { instance } from "../../../utils/interceptor";
+import { MODAL_ROOT } from "../../../utils";
 import AmountInput from "../../../components/AmountInput";
 import InputField from "../../../components/InputField";
 import LoadingSpinner from "../../../components/LoadingSpinner/LoadingSpinner";
