+++ src/pages/Home/TransactionCard/TransactionDetailModal/TransactionDetailModal.tsx
@@ -3,7 +3,7 @@ import { createPortal } from "react-dom";
 import { useTranslation } from "react-i18next";
 import ModalDialog from "../../../../layouts/ModalDialog";
 import { Transaction } from "../../../../models/transaction.model";
-import { MODAL_ROOT } from "../../../../util";
+import { MODAL_ROOT } from "../../../../utils";
 import LNDetails from "./LNDetails";
 import OnchainDetails from "./OnchainDetails";
