+++ src/pages/Home/index.tsx
@@ -7,9 +7,9 @@ import { useInterval } from "../../hooks/use-interval";
 import useSSE from "../../hooks/use-sse";
 import { AppStatus } from "../../models/app-status";
 import { Transaction } from "../../models/transaction.model";
-import { checkError } from "../../util/checkError";
-import { instance } from "../../util/interceptor";
-import { enableGutter } from "../../util";
+import { checkError } from "../../utils/checkError";
+import { instance } from "../../utils/interceptor";
+import { enableGutter } from "../../utils";
 import AppStatusCard from "./AppStatusCard";
 import BitcoinCard from "./BitcoinCard";
 import ConnectionCard from "./ConnectionCard";
