+++ src/utils/__tests__/checkError.test.ts
