+++ src/utils/availableApps.ts
